API: Serde bindings are now in separate crate.
