+++ rmp-serde/src/lib.rs
@@ -1,0 +1,8 @@
+extern crate rmp;
+extern crate serde;
+
+pub mod decode;
+pub mod encode;
+
+pub use decode::Deserializer;
+pub use encode::Serializer;
