+++ rmp/tests/func/decode/mod.rs
@@ -10,6 +10,5 @@ mod map;
 mod ext;
 
 // High-level deserialize interface.
-mod deserializer;
 mod value;
 mod value_ref;
