+++ rmp/tests/func/mod.rs
@@ -1,3 +1,2 @@
 mod encode;
 mod decode;
-mod serde;
