+++ rmp/tests/lib.rs
@@ -1,7 +1,3 @@
-#![cfg_attr(feature = "serde_macros", feature(custom_derive, plugin))]
-#![cfg_attr(feature = "serde_macros", plugin(serde_macros))]
-
 extern crate rmp as msgpack;
-extern crate serde;
 
 mod func;
